added keyring to store keys, but it's not stable yet

+++ build-linux/install.sh
@@ -1,19 +1,47 @@
 #!/bin/bash
 
-# Define the URLs for the files :: this is a test scr
 URL="https://repo.dhanu.cloud/clippy/clippy-release.tar.xz"
 ARCHIVE_NAME="clippy-release.tar.xz"
 TARGET_DIR="clippy"
+SOURCE_DIR="target/release"
+BUILD_LINUX="build-linux"
+BUILD_ASSETS="assets"
 
-echo "Downloading file..."
-curl -L -o "$ARCHIVE_NAME" "$URL"
-mkdir -p "$TARGET_DIR"
-tar -xJf "$ARCHIVE_NAME" -C "$TARGET_DIR"
-cd "$TARGET_DIR" || { echo "Failed to cd into $TARGET_DIR"; exit 1; }
+do_r=false
+do_d=false
+
+while getopts "rd" opt; do
+  case $opt in
+  r) do_r=true ;;
+  d) do_d=true ;;
+  \?)
+    echo "Invalid option: -$OPTARG" >&2
+    exit 1
+    ;;
+  esac
+done
+if $do_r; then
+  SOURCE_DIR="../target/release"
+  BUILD_LINUX="../build-linux"
+  BUILD_ASSETS="../assets"
+elif $do_d; then
+  SOURCE_DIR="../target/debug"
+  BUILD_LINUX="../build-linux"
+  BUILD_ASSETS="../assets"
+else
+  echo "Downloading file..."
+  curl -L -o "$ARCHIVE_NAME" "$URL"
+  mkdir -p "$TARGET_DIR"
+  tar -xJf "$ARCHIVE_NAME" -C "$TARGET_DIR"
+  cd "$TARGET_DIR" || {
+    echo "Failed to cd into $TARGET_DIR"
+    exit 1
+  }
+fi
 
 # Define the target locations
-BIN_DIR="/usr/local/bin"  
-SERVICE_DIR="/etc/systemd/user"   
+BIN_DIR="/usr/local/bin"
+SERVICE_DIR="/etc/systemd/user"
 SERVICE_NAME="clippy.service"
 
 systemctl --user stop "$SERVICE_NAME"
@@ -22,25 +50,32 @@ echo "Copying file..."
 # Create the service directory if it doesn't exist
 mkdir -p "$SERVICE_DIR"
 
-sudo cp  "target/release/clippy"  "$BIN_DIR/clippy"
-sudo cp  "target/release/clippy-gui"  "$BIN_DIR/clippy-gui"
+sudo cp "$SOURCE_DIR/clippy" "$BIN_DIR/clippy"
+sudo cp "$SOURCE_DIR/clippy-gui" "$BIN_DIR/clippy-gui"
 
-sudo cp "build-linux/clippy.service" "$SERVICE_DIR/$SERVICE_NAME"
+sudo cp "$BUILD_LINUX/clippy.service" "$SERVICE_DIR/$SERVICE_NAME"
 
-for f in assets/icons/clippy-*.png; do
+for f in "$ASSTS"/icons/clippy-*.png; do
+  [[ -e "$f" ]] || continue # skip if no file matched
+
   base=$(basename "$f")
-  size=$(echo "$base" | sed -E 's/clippy-([0-9]+)-([0-9]+)(@2)?\.png/\1x\2\3/')
+  size=$(echo "$base" | sed -E 's/clippy-([0-9]+)-([0-9]+)(@2)?\.png/\1x\2/')
   folder=$(echo "$size" | sed 's/@2//')
   target="/usr/share/icons/hicolor/${folder}/apps"
+
   sudo mkdir -p "$target"
   sudo cp "$f" "$target/clippy.png"
   echo "Copied $f -> $target/clippy.png"
 done
 
-sudo cp build-linux/clippy.desktop /usr/share/applications/clippy.desktop
+sudo cp "$BUILD_LINUX/clippy.desktop" "/usr/share/applications/clippy.desktop"
 
-rm -r "&TARGET_DIR"
-rm "&ARCHIVE_NAME"
+if ! $do_d && ! $do_r; then
+  echo "Deleting temp files..."
+
+  rm -r "$TARGET_DIR"
+  rm "$ARCHIVE_NAME"
+fi
 
 # Reload systemd, enable, and start the service
 echo "Enabling and starting the service..."
